move custom hook to utils file
