+++ packages/react-composites/src/composites/common/AvatarPersona.tsx
@@ -3,6 +3,7 @@
 
 import { IPersonaProps, Persona, PersonaInitialsColor } from '@fluentui/react';
 import React, { useEffect, useMemo } from 'react';
+import { useCustomAvatarPersonaData } from '../CallComposite/utils';
 
 /**
  * Custom data attributes for displaying avatar for a user.
@@ -81,76 +82,4 @@ export const AvatarPersona = (props: AvatarPersonaProps): JSX.Element => {
       initialsTextColor={data?.initialsTextColor ?? initialsTextColor ?? 'white'}
     />
   );
-};
-
-/**
- * Hook to override the avatar persona data of the users
- * - Calls {@link AvatarPersonaDataCallback} on each render for provided `userIds`
- * - Returns an array of the same length as `userIds`. Entries in the array may be undefined if there is no data to return.
- * @private
- */
-export const useCustomAvatarPersonaData = (
-  // move this to utils
-  userIds: (string | undefined)[],
-  callBack?: AvatarPersonaDataCallback
-): (AvatarPersonaData | undefined)[] => {
-  const [data, setData] = React.useState<(AvatarPersonaData | undefined)[]>([]);
-  useEffect(() => {
-    (async () => {
-      if (callBack) {
-        const newData = await Promise.all(
-          userIds.map(async (userId: string | undefined) => {
-            if (!userId) {
-              return undefined;
-            }
-            return await callBack(userId);
-          })
-        );
-        if (shouldUpdate(data, newData)) {
-          setData(newData);
-        }
-      }
-    })();
-  });
-  return data;
-};
-
-/**
- * Function to determine if there is new user avatar data in current render pass.
- * @param currentData current set set of avatar persona data present from the custom settings from the previous render
- * @param newData new set of avatar persona data after a run of {@link useCustomAvatarPersonData}
- * @returns Boolean whether there is new avatar persona data present.
- * @private
- */
-const shouldUpdate = (
-  currentData: (AvatarPersonaData | undefined)[],
-  newData: (AvatarPersonaData | undefined)[]
-): boolean => {
-  let newDataPresent = false;
-  if (currentData.length !== newData.length) {
-    return true;
-  }
-  newDataPresent = avatarDeepEqual(currentData, newData);
-  return newDataPresent;
-};
-
-/**
- * @private
- */
-const avatarDeepEqual = (
-  currentData: (AvatarPersonaData | undefined)[],
-  newData: (AvatarPersonaData | undefined)[]
-): boolean => {
-  let newDataPresent;
-  currentData.forEach((p, i) => {
-    newDataPresent =
-      p?.text !== newData[i]?.text &&
-      p?.imageUrl !== newData[i]?.imageUrl &&
-      p?.initialsColor !== newData[i]?.initialsColor &&
-      p?.imageInitials !== newData[i]?.imageInitials &&
-      p?.initialsTextColor !== newData[i]?.initialsTextColor
-        ? true
-        : false;
-  });
-  return newDataPresent;
 };
